Weapon Balance
Quality of life tweaks to different weapons, main focus is not having massive differences between guns that are quite similar to each other - especially when firing the same calibre rounds. Weapons that have a change are the more expensive, often less durable modern alternatives which may have less recoil and spread in exchange for durability / cost.

Also the SMG's had a bit of a weird gap between them stat wise, even as weapons that are pretty comparable to each other.