Bullet Changes
Major change is that the 7.62 AK round is no longer comparable to sniper rifles in damage, which it previously was and isn't really something it's meant for. If you are curious you can look at some ballistics to compare 5.56 to 7.62, for the most part there isn't like a crazy difference and 5.56 seems to outperform it in the most important ways. Some tests even show it does better for soft tissue damage because it tumbles inside the body and causes more damage.

Also outside of the nerd bullet shiz, most of the ammo is like the guns - comparable to each other and a bit better balanced towards realism. I did leave a little bit of the hollywood'ness of the .44 magnum so it remains a high damage weapon but that's mostly because the desert eagle and anaconda are really expensive to buy. .44 in ballistics tests isn't very scary compared to 5.56, especially when fired from a short barrel.